Created a static HTML page for testing purposes
One reason why we are creating the <clipspace-article>
element is because we want to make the development easier.

The advantage of polymer is that we can work on the polymer
emlement without having to start all the server dependencies.

All we need for now is a static age where we can place the
new element and work on it.

In this commit I have just used chrome to download the page
frome the portal to have static page. It still has some issue
and I hope you guys will fix it.