Imported CSS
I have just imported the generated CSS
from the workspace page. However, since
this is generated CSS we need to find a
decent way to import these files into
the polymer element.

There is are still some styling issue
that have to be resolved.

Futhermore we need to make sure that
we can style the element without having
to compile the element.

This could be an entrypoint:
https://www.polymer-project.org/1.0/start/first-element/step-5