Introduce property 'date'
This property accepts a prerendered
String of a date.

Usually it would make sense to pass
a date value and let it render by
the component. However we have configuration
option in the backend where we can descide
if the date is suppose to be shown with or
without the daytime.

Since I don't want to make it complicated
right now I'll leave it as simple string
for now.