Introduced i18next
The application is multilingual and
therefore we need a decent mechanism
to handle translations. I have choosen
i18next which appears quite mature to
me.

I have currently just stubbed the translation.

TODO: Maintain a translation file. In order
to avoid redundancies I think it would be the
best to create a central file for all the
translations that are handled in the frontend.

Therefore the XHR loading must be enabled in
i18next. The respective module is already
loaded in this commit.